db/sqlc/store.go
package db

import (
	"context"
	"database/sql"
	"fmt"
)

// Store用于提供所有的func执行db的查询和交易
type Store struct {
	*Queries
	db *sql.DB
}

func NewStore(db *sql.DB) *Store {
	return &Store{
		db:      db,
		Queries: New(db),
	}
}

//测试
//var txKey = struct{}{}

// 执行数据库事务用于交易
// 需要一个上下文和回调函数作为输入
// 小写e开头，无法导出
func (store *Store) execTx(ctx context.Context, fn func(*Queries) error) error {
	//BeginTx会返回事务对象或错误
	tx, err := store.db.BeginTx(ctx, nil)
	if err != nil {
		return err
	}
	q := New(tx)
	err = fn(q)
	if err != nil {
		//如果事务查询错误且回滚错误
		if rbErr := tx.Rollback(); rbErr != nil {

			return fmt.Errorf("tx err: %v,rb err: %v", err, rbErr)
		}
		return err
	}
	return tx.Commit()
}

// TransfersTxParams包含在两个账户交易之间的输入参数
type TransferTxParams struct {
	FromAccountID int64 `json:"from_account_id"`
	ToAccountID   int64 `json:"to_account_id"`
	Amount        int64 `json:"amount"`
}

// TransfersTxResults结构包含传输事务的结果
type TransferTxResult struct {
	Transfer    Transfers `json:"transfer"`
	FromAccount Accounts  `json:"from_account"`
	ToAccount   Accounts  `json:"to_account"`
	FromEntry   Entries   `json:"from_entry"`
	ToEntry     Entries   `json:"to_entry"`
}

// TransfersTx执行交易实例
// 创建一个新的传输记录，添加新的账户条目，并更新账户余额在单个数据库事务中
func (store *Store) TransferTx(ctx context.Context, arg TransferTxParams) (TransferTxResult, error) {
	var result TransferTxResult

	err := store.execTx(ctx, func(q *Queries) error {
		//回调函数
		var err error
		//测试
		//txName := ctx.Value(txKey)
		//fmt.Println(txName, "create transfer")
		result.Transfer, err = q.CreateTransfer(ctx, CreateTransferParams{
			FromAccountID: arg.FromAccountID,
			ToAccountID:   arg.ToAccountID,
			Amount:        arg.Amount,
		})
		if err != nil {
			return err
		}
		//回调函数不知道应该返回结果的确切类型，使用闭包

		//fmt.Println(txName, "create Entry 1")
		result.FromEntry, err = q.CreateEntry(ctx, CreateEntryParams{
			AccountID: arg.FromAccountID,
			Amount:    -arg.Amount,
		})
		if err != nil {
			return err
		}

		//fmt.Println(txName, "create Entry 2")
		result.ToEntry, err = q.CreateEntry(ctx, CreateEntryParams{
			AccountID: arg.ToAccountID,
			Amount:    arg.Amount,
		})
		if err != nil {
			return err
		}
		//更新账户余额，小心处理并发避免死锁
		//q.GetAccount(获取from_account记录并将其分配给account1变量
		//fmt.Println(txName, "get account 1")

		//fmt.Println(txName, "update account 1")

		// 防止死锁最好的方法是确保程序总是以一致的顺序获取锁
		if arg.FromAccountID < arg.ToAccountID {
			result.FromAccount, result.ToAccount, err = addMoney(ctx, q, arg.FromAccountID, -arg.Amount, arg.ToAccountID, arg.Amount)
		} else {
			result.ToAccount, result.FromAccount, err = addMoney(ctx, q, arg.ToAccountID, arg.Amount, arg.FromAccountID, -arg.Amount)
		}
		return nil
	})

	return result, err
}

func addMoney(
	ctx context.Context,
	q *Queries,
	accountID1 int64,
	amount1 int64,
	accountID2 int64,
	amount2 int64,
) (account1 Accounts, account2 Accounts, err error) {
	//调用q.AddAccountBalance()将金额1添加到account1，
	account1, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
		ID:     accountID1,
		Amount: amount1,
	})
	if err != nil {
		return //使用了命名返回变量，简洁处理直接return
	}

	account2, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
		ID:     accountID2,
		Amount: amount2,
	})
	return

}
